Added request responses to the README, Rounded the return values of the api

+++ src/main/kotlin/features/accounting/model/BenfordResult.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class BenfordResult(
-    val observedDistribution: Map<Int, Double>,  // Actual percentages
-    val expectedDistribution: Map<Int, Double>, // Benford's Law theoretical percentages
+    val observedDistribution: Map<Int, Int>,  // Actual percentages
+    val expectedDistribution: Map<Int, Int>, // Benford's Law theoretical percentages
     val chiSquareStatistic: Double,
     val criticalValue: Double,
     val isBenfordCompliant: Boolean
